pyproject.toml
[project]
name = "web-chat-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "sqlalchemy[asyncio]>=2.0.40",
    "uvicorn>=0.34.2",
    "yoyo-migrations>=9.0.0",
]
